refactor: set esm to libs
