+++ libs/ansi-parser/package.json
@@ -1,6 +1,7 @@
 {
   "name": "ansi-parser",
   "version": "0.0.0",
+  "type": "module",
   "description": "",
   "private": true,
   "keywords": [
