+++ libs/ansi-parser/src/index.ts
@@ -1,7 +1,7 @@
-export { default as colorIndexOf } from './color-index-of';
-export { default as decodeAnsiByte } from './decode-ansi-byte';
-export { default as decodeAnsiBytes } from './decode-ansi-bytes';
-export { default as decodeAnsiColor8bit } from './decode-ansi-color-8bit';
-export { default as decodeAnsiString } from './decode-ansi-string';
-export { default as defaultAnsiParser } from './default-ansi-parser';
-export { default as encodeAnsiAction } from './encode-ansi-action';
+export { default as colorIndexOf } from './color-index-of.js';
+export { default as decodeAnsiByte } from './decode-ansi-byte.js';
+export { default as decodeAnsiBytes } from './decode-ansi-bytes.js';
+export { default as decodeAnsiColor8bit } from './decode-ansi-color-8bit.js';
+export { default as decodeAnsiString } from './decode-ansi-string.js';
+export { default as defaultAnsiParser } from './default-ansi-parser.js';
+export { default as encodeAnsiAction } from './encode-ansi-action.js';
