+++ libs/notios-config/tsconfig.json
@@ -1,19 +1,13 @@
 {
   "compilerOptions": {
     "esModuleInterop": true,
-    "module": "CommonJS",
-    "moduleResolution": "Node",
+    "module": "Node16",
+    "moduleResolution": "Node16",
     "skipLibCheck": true,
     "strict": true,
     "target": "ES2015",
     "incremental": true
   },
-  "exclude": [
-    "node_modules",
-    "test"
-  ],
-  "include": [
-    "**/*.ts",
-    "**/*.tsx"
-  ]
+  "exclude": ["node_modules", "test"],
+  "include": ["**/*.ts", "**/*.tsx"]
 }
