+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "notios",
-  "type": "module",
   "version": "0.4.1",
+  "type": "module",
   "description": "",
   "keywords": [
     "notios",
